+++ src/Anexia.MathematicalProgram/SolverConfiguration/LPSolverType.cs
@@ -27,5 +27,5 @@ public enum LpSolverType
     /// <see href="https://gurobi.com">Gurobi</see> solver. A licence is needed for usage.
     /// </summary>
     [EnumMember(Value = "GUROBI_MIXED_INTEGER_PROGRAMMING")]
-    GurobiIntegerProgramming,
+    GurobiMixedIntegerProgramming,
 }