+++ CHANGELOG.md
@@ -7,6 +7,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.2.0] - 2025-03-26
+
+### Added
+
+- Added possibility to pass an MPS model as string to the LP solver.
+- Added possibility to pass additional solver specific parameters to the solvers.
+- Added public members in ICompletedOptimizationModel.
+- Added class IntegralPoint for integral intervals.
+
+### Fixed
+
+- Fixed bug where the GLOP solver could not be used with NumberOfThreads parameter.
+
 ## [2.1.0] - 2025-03-05
 
 ### Added